requirements.txt
uvicorn==0.29.0
fastapi==0.110.0
pydantic==2.6.4
python-dotenv==1.0.1
pydantic-settings==2.2.1